+++ src/main/java/be/baur/sda/io/SDAFormatter.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package be.baur.sda.serialization;
+package be.baur.sda.io;
 
 import java.io.IOException;
 import java.io.Writer;