+++ src/test/java/samples/parser/SDAParser.java
@@ -6,8 +6,8 @@ import java.io.Reader;
 import java.util.Stack;
 
 import be.baur.sda.DataNode;
-import be.baur.sda.serialization.Parser;
-import be.baur.sda.serialization.SDAParseException;
+import be.baur.sda.io.Parser;
+import be.baur.sda.io.SDAParseException;
 
 /**
  * Alternative SDA parser (actually the first one I wrote, and lacking support