+++ RELEASE_NOTES.md
@@ -12,6 +12,7 @@ In development.
 - Added convenience methods SDA.format(File, DataNode) and SDA.format(DataNode).
 - Closed issue #16 (SDAFormatter throws if input node is null).
 - Closed issue #12 (Add a parse(File) method and review exceptions).
+- Closed issue #17 (Inappropriate IllegalStateException in AbstractNode.add()).
 
 ## Compatibility